requirements.txt
# Install dependencies
# libreoffice
# poppler
# ffmpeg
pdf2image
python-pptx
openai
requests